layouts/_default/baseof.html: Move masthead, header and footer into partial templates
I'd like to override some aspects of how the masthead works
(specifically get rid of the 'Home' menu item and re-implement the rules
for when a navbar item is active). I think the only way to do this
currently is to override the entire baseof.html file.

By moving these three pieces into a partial, it allows more control for
overriding and makes it easier to keep the theme up-to-date.