Update to Font Awesome 5
Nothing really new as far as how we're using Font Awesome, but it is
good to keep up with the times and the tooling. Users that have used
custom fonts in their content or layouts will need to update their
icon prefixes from "fa" to "fas" or "fab" depending on which font
their icon comes from.

See: https://fontawesome.com/how-to-use/on-the-web/setup/upgrading-from-version-4