.travis.yml: Use an LXD container instead of VM
We need to specify OS and Arch in order to use an LXD container in-
stead of a VM. This should theoretically be much faster according
to the Travis docs. Here we are building on arm64 because there is
no amd64 container support right now, but it shouldn't matter beca-
use we are only doing npm build.

See: https://docs.travis-ci.com/user/reference/overview/